refactor: Move test_extract_hyperlinks to test_browse_scrape_links.py tests
